Make the number of relative line numbers shown customizable
I don't know why I want this feature.. but I do.

If I was going to create a justification after the fact
it would be that things changing away from the input
point is distracting.

This feature also allows you to see the "relative insertion locations"
as moving together with the cursor